+++ ReleaseNotes.md
@@ -1,6 +1,8 @@
-### New in 0.0.4 (Released xxxx/xx/xx)
+### New in 0.0.4 (Released 2014/12/10)
 * Added patching of .gitignore.
 * Added AppVeyor configuration file.
+* Replaced Start-Process with Invoke-Expression in bootstrapper script template.
+* Now returns exit code from bootstrapper script template.
 
 ### New in 0.0.3 (Released 2014/11/29)
 * Added default build script template functionality.